+++ scripts/sift-worker.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+import os
+import sys
+sys.path.append(os.path.dirname(os.path.abspath(os.path.join(__file__, '..'))))
+
+from sift import main
+
+if __name__ == '__main__':
+    main()
